requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
confection==0.1.5
cymem==2.0.8
distro==1.9.0
Django==4.2
djangorestframework==3.15.2
drf-yasg==1.21.7
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
frozenlist==1.4.1
graphene==3.3
graphene-django==3.2.2
graphql-core==3.2.3
graphql-relay==3.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inflection==0.5.1
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
murmurhash==1.0.10
numpy==1.26.4
packaging==24.1
pandas==2.2.2
preshed==3.0.9
promise==2.3
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.1
srsly==2.4.8
text-unidecode==1.3
thinc==8.2.5
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
wasabi==1.1.3
weasel==0.4.1
wrapt==1.16.0
yarl==1.9.4
